chore(ts): migrate ImageSource, RasterDemSource, RasterSource to typescript
