+++ docs/ImageSource.md
@@ -1,4 +1,4 @@
-<!-- This file was autogenerated from ImageSource.js do not modify -->
+<!-- This file was autogenerated from ImageSource.tsx do not modify -->
 # <MapboxGL.ImageSource />
 ImageSource is a content source that is used for a georeferenced raster image to be shown on the map.
 The georeferenced image scales and rotates as the user zooms and rotates the map
@@ -7,7 +7,8 @@ The georeferenced image scales and rotates as the user zooms and rotates the map
 | Prop | Type | Default | Required | Description |
 | ---- | :-- | :----- | :------ | :---------- |
 | id | `string` | `none` | `true` | A string that uniquely identifies the source. |
-| url | `union` | `none` | `false` | An HTTP(S) URL, absolute file URL, or local file URL to the source image.<br/>Gifs are currently not supported. |
-| coordinates | `array` | `none` | `true` | The top left, top right, bottom right, and bottom left coordinates for the image. |
+| url | `number \| string` | `none` | `false` | An HTTP(S) URL, absolute file URL, or local file URL to the source image.<br/>Gifs are currently not supported. |
+| coordinates | `tuple` | `none` | `true` | The top left, top right, bottom right, and bottom left coordinates for the image. |
+| children | `React.ReactElement \| React.ReactElement[]` | `none` | `false` | FIX ME NO DESCRIPTION |
